requirements.txt
streamlit
pandas
requests
python-dotenv
langchain-google-genai
